Moved dictionary extension to its own file. Added test
